pjt08/requirements.txt
asgiref==3.4.1
backcall==0.2.0
colorama==0.4.4
decorator==5.1.0
Django==3.2.7
django-extensions==3.1.3
djangorestframework==3.12.4
ipython==7.27.0
jedi==0.18.0
matplotlib-inline==0.1.3
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.20
Pygments==2.10.0
pytz==2021.1
sqlparse==0.4.1
traitlets==5.1.0
wcwidth==0.2.5
